Support preflight OPTIONS requests for ajax
